requirements.txt
boto==2.48.0
boto3==1.7.55
botocore==1.10.55
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
cycler==0.10.0
decorator==4.3.0
docutils==0.14
flake8==3.5.0
gensim==3.5.0
idna==2.7
jmespath==0.9.3
JPype1==0.6.3
kiwisolver==1.0.1
matplotlib==2.2.2
mccabe==0.6.1
networkx==2.1
nltk==3.3
numpy==1.14.5
py-dateutil==2.2
pybuilder==0.11.17
pycodestyle==2.3.1
pyflakes==1.6.0
pymongo==3.7.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
smart-open==1.6.0
sutime==0.2.0
tailer==0.4.1
tblib==1.3.2
urllib3==1.23
waybackpack==0.3.5
